Move visitor classes to visitor package
