+++ src/main/java/datawave/data/normalizer/regex/AnyCharNode.java
@@ -1,5 +1,7 @@
 package datawave.data.normalizer.regex;
 
+import datawave.data.normalizer.regex.visitor.Visitor;
+
 /**
  * Represents a dot in a regex pattern.
  */
