+++ src/main/java/datawave/data/normalizer/regex/CharClassNode.java
@@ -1,5 +1,7 @@
 package datawave.data.normalizer.regex;
 
+import datawave.data.normalizer.regex.visitor.Visitor;
+
 import java.util.Objects;
 
 /**
