+++ src/main/java/datawave/data/normalizer/regex/EmptyNode.java
@@ -1,5 +1,7 @@
 package datawave.data.normalizer.regex;
 
+import datawave.data.normalizer.regex.visitor.Visitor;
+
 /**
  * Placeholder empty node for empty groups or empty alternation branches.
  */
