+++ src/main/java/datawave/data/normalizer/regex/ExpressionNode.java
@@ -1,5 +1,7 @@
 package datawave.data.normalizer.regex;
 
+import datawave.data.normalizer.regex.visitor.Visitor;
+
 import java.util.List;
 
 /**
