+++ src/main/java/datawave/data/normalizer/regex/OptionalNode.java
@@ -1,5 +1,7 @@
 package datawave.data.normalizer.regex;
 
+import datawave.data.normalizer.regex.visitor.Visitor;
+
 /**
  * Represents the question mark in a regex pattern.
  */
