+++ src/main/java/datawave/data/normalizer/regex/RepetitionNode.java
@@ -1,5 +1,7 @@
 package datawave.data.normalizer.regex;
 
+import datawave.data.normalizer.regex.visitor.Visitor;
+
 /**
  * Represents a repetition requirement in a regex pattern, e.g. {@code {3}}.
  */
