+++ src/main/java/datawave/data/normalizer/regex/ZeroOrMoreNode.java
@@ -1,5 +1,7 @@
 package datawave.data.normalizer.regex;
 
+import datawave.data.normalizer.regex.visitor.Visitor;
+
 /**
  * Represents the star in a regex pattern.
  */
