+++ src/main/java/datawave/data/normalizer/regex/visitor/BinFinder.java
@@ -1,5 +1,11 @@
-package datawave.data.normalizer.regex;
+package datawave.data.normalizer.regex.visitor;
 
+import datawave.data.normalizer.regex.IntegerNode;
+import datawave.data.normalizer.regex.IntegerRangeNode;
+import datawave.data.normalizer.regex.Node;
+import datawave.data.normalizer.regex.NodeListIterator;
+import datawave.data.normalizer.regex.RegexConstants;
+import datawave.data.normalizer.regex.RegexUtils;
 import org.apache.commons.lang3.tuple.Pair;
 
 /**
