+++ src/main/java/datawave/data/normalizer/regex/visitor/GTEOneBinFinder.java
@@ -1,6 +1,7 @@
-package datawave.data.normalizer.regex;
+package datawave.data.normalizer.regex.visitor;
 
-import datawave.data.normalizer.regex.visitor.StringVisitor;
+import datawave.data.normalizer.regex.Node;
+import datawave.data.normalizer.regex.RegexUtils;
 import org.apache.commons.lang3.tuple.Pair;
 
 /**
