+++ src/main/java/datawave/data/normalizer/regex/visitor/PrintVisitor.java
@@ -20,7 +20,6 @@ import datawave.data.normalizer.regex.OptionalNode;
 import datawave.data.normalizer.regex.RepetitionNode;
 import datawave.data.normalizer.regex.SingleCharNode;
 import datawave.data.normalizer.regex.StartAnchorNode;
-import datawave.data.normalizer.regex.Visitor;
 import datawave.data.normalizer.regex.ZeroOrMoreNode;
 
 /**
