+++ src/main/java/datawave/data/normalizer/regex/visitor/Visitor.java
@@ -1,4 +1,25 @@
-package datawave.data.normalizer.regex;
+package datawave.data.normalizer.regex.visitor;
+
+import datawave.data.normalizer.regex.AlternationNode;
+import datawave.data.normalizer.regex.AnyCharNode;
+import datawave.data.normalizer.regex.CharClassNode;
+import datawave.data.normalizer.regex.CharRangeNode;
+import datawave.data.normalizer.regex.DigitCharClassNode;
+import datawave.data.normalizer.regex.EmptyNode;
+import datawave.data.normalizer.regex.EncodedNumberNode;
+import datawave.data.normalizer.regex.EncodedPatternNode;
+import datawave.data.normalizer.regex.EndAnchorNode;
+import datawave.data.normalizer.regex.EscapedSingleCharNode;
+import datawave.data.normalizer.regex.ExpressionNode;
+import datawave.data.normalizer.regex.GroupNode;
+import datawave.data.normalizer.regex.IntegerNode;
+import datawave.data.normalizer.regex.IntegerRangeNode;
+import datawave.data.normalizer.regex.OneOrMoreNode;
+import datawave.data.normalizer.regex.OptionalNode;
+import datawave.data.normalizer.regex.RepetitionNode;
+import datawave.data.normalizer.regex.SingleCharNode;
+import datawave.data.normalizer.regex.StartAnchorNode;
+import datawave.data.normalizer.regex.ZeroOrMoreNode;
 
 public interface Visitor {
